Converted legacy services to routers
