+++ nodeServer/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "org.zowe.zosmf.workflows.nodeserver",
+  "version": "0.9.5",
+  "description": "Backend communication with z/OSMF",
+  "scripts": {
+    "build": "tsc",
+    "start": "tsc --watch"
+  },
+  "dependencies": {},
+  "devDependencies": {
+    "bluebird": "~3.5.1",
+    "body-parser": "~1.18.3",
+    "express": "~4.16.3",
+    "@types/express": "~4.16.0",
+    "@types/node": "~8.10.23",
+    "typescript": "~2.7.2"
+  }
+}
