+++ nodeServer/tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compileOnSave": true,
+  "include": [
+    "ts/*.ts"
+  ],
+  "compilerOptions": {
+    "target": "es2015",
+    "outDir": "../lib",
+    "sourceMap": true,
+    "moduleResolution": "node",
+    "types": [
+      "node",
+      "express"
+    ],
+    "lib": [
+      "es5",
+      "es6"
+    ]
+  }
+}
